# Conflicts:
#	Accounting/Configurations/MappingProfile.cs
#	Accounting/Context/DapperContext.cs
#	Accounting/Controllers/APInvoicesController.cs
#	Accounting/Controllers/ARInvoicesController.cs
#	Accounting/Controllers/BusinessPartnersController.cs
#	Accounting/Controllers/DepreciationsController.cs
#	Accounting/Controllers/ExpenseCategoriesController.cs
#	Accounting/Controllers/FixedAssetsController.cs
#	Accounting/Controllers/LoansController.cs
#	Accounting/Controllers/PeriodsController.cs
#	Accounting/Controllers/TenantsController.cs
#	Accounting/DTOs/APInvoiceDTO.cs
#	Accounting/DTOs/APInvoiceLineDTO.cs
#	Accounting/DTOs/ARInvoiceDTO.cs
#	Accounting/DTOs/ARInvoiceLineDTO.cs
#	Accounting/DTOs/BusinessPartnerDTO.cs
#	Accounting/DTOs/CreateAPInvoiceDTO.cs
#	Accounting/DTOs/CreateAPInvoiceLineDTO.cs
#	Accounting/DTOs/CreateARInvoiceDTO.cs
#	Accounting/DTOs/CreateARInvoiceLineDTO.cs
#	Accounting/DTOs/CreateBusinessPartnerDTO.cs
#	Accounting/DTOs/CreateDepreciationDTO.cs
#	Accounting/DTOs/CreateExpenseCategoryDTO.cs
#	Accounting/DTOs/CreateFixedAssetDTO.cs
#	Accounting/DTOs/CreateLoanDTO.cs
#	Accounting/DTOs/CreateTenantDTO.cs
#	Accounting/DTOs/DepreciationDTO.cs
#	Accounting/DTOs/ExpenseCategoryDTO.cs
#	Accounting/DTOs/FixedAssetDTO.cs
#	Accounting/DTOs/FixedAssetYearDetailsDTO.cs
#	Accounting/DTOs/LoanDTO.cs
#	Accounting/DTOs/PeriodDTO.cs
#	Accounting/DTOs/TenantDTO.cs
#	Accounting/DTOs/UpdateAPInvoiceDTO.cs
#	Accounting/DTOs/UpdateAPInvoiceLineDTO.cs
#	Accounting/DTOs/UpdateARInvoiceDTO.cs
#	Accounting/DTOs/UpdateARInvoiceLineDTO.cs
#	Accounting/DTOs/UpdateDepreciationDTO.cs
#	Accounting/DTOs/UpdateFixedAssetDTO.cs
#	Accounting/Middlewares/GlobalErrorHandlingMiddleware.cs
#	Accounting/Models/APInvoice.cs
#	Accounting/Models/APInvoiceLine.cs
#	Accounting/Models/ARInvoice.cs
#	Accounting/Models/ARInvoiceLine.cs
#	Accounting/Models/BaseModel.cs
#	Accounting/Models/BusinessPartner.cs
#	Accounting/Models/Depreciation.cs
#	Accounting/Models/ExpenseCategory.cs
#	Accounting/Models/FixedAsset.cs
#	Accounting/Models/Invoice.cs
#	Accounting/Models/InvoiceLine.cs
#	Accounting/Models/Loan.cs
#	Accounting/Models/Period.cs
#	Accounting/Models/Tenant.cs
#	Accounting/Program.cs
#	Accounting/Repositories/APInvoiceLinesRepository.cs
#	Accounting/Repositories/APInvoiceRepository.cs
#	Accounting/Repositories/ARInvoiceLinesRepository.cs
#	Accounting/Repositories/ARInvoiceRepository.cs
#	Accounting/Repositories/BusinessPartnerRepository.cs
#	Accounting/Repositories/DepreciationRepository.cs
#	Accounting/Repositories/ExpenseCategoryRepository.cs
#	Accounting/Repositories/FixedAssetRepository.cs
#	Accounting/Repositories/IAPInvoiceLineRepository.cs
#	Accounting/Repositories/IAPInvoiceRepository.cs
#	Accounting/Repositories/IARInvoiceLineRepository.cs
#	Accounting/Repositories/IARInvoiceRepository.cs
#	Accounting/Repositories/IBusinessPartnerRepository.cs
#	Accounting/Repositories/IDepreciationRepository.cs
#	Accounting/Repositories/IExpenseCategoryRepository.cs
#	Accounting/Repositories/IFixedAssetRepository.cs
#	Accounting/Repositories/ILoanRepository.cs
#	Accounting/Repositories/IPeriodRepository.cs
#	Accounting/Repositories/ITenantRepository.cs
#	Accounting/Repositories/LoanRepository.cs
#	Accounting/Repositories/PeriodRepository.cs
#	Accounting/Repositories/TenantRepository.cs
#	Accounting/Services/APInvoiceService.cs
#	Accounting/Services/ARInvoiceService.cs
#	Accounting/Services/BusinessPartnerService.cs
#	Accounting/Services/DepreciationService.cs
#	Accounting/Services/ExpenseCategoryService.cs
#	Accounting/Services/FixedAssetService.cs
#	Accounting/Services/IAPInvoiceService.cs
#	Accounting/Services/IARInvoiceService.cs
#	Accounting/Services/IBusinessPartnerService.cs
#	Accounting/Services/IDepreciationService.cs
#	Accounting/Services/IExpenseCategoryService.cs
#	Accounting/Services/IFixedAssetService.cs
#	Accounting/Services/ILoanService.cs
#	Accounting/Services/IPeriodService.cs
#	Accounting/Services/ITenantService.cs
#	Accounting/Services/LoanService.cs
#	Accounting/Services/PeriodService.cs
#	Accounting/Services/TenantService.cs
#	Accounting/Validators/CreateAPInvoiceDTOValidator.cs
#	Accounting/Validators/CreateARInvoiceDTOValidator.cs
#	Accounting/Validators/CreateExpenseCategoryDTOValidator.cs
#	Accounting/Validators/CreatePeriodDTOValidator.cs
#	Accounting/Validators/CreateTenantDTOValidator.cs
#	Accounting/Validators/UpdateAPInvoiceDTOValidator.cs
#	Accounting/Validators/UpdateARInvoiceDTOValidator.cs
#	Auth/Program.cs
#	Auth/appsettings.Production.json
#	ContactsAPI/ContactsAPI.csproj
#	ContactsAPI/Program.cs
#	Documents/DocumentsAPI.csproj
#	OcelotApiGw/Program.cs
#	OcelotApiGw/Properties/launchSettings.json
#	OcelotApiGw/ocelot.json

+++ Documents/Controllers/DocumentsController.cs
@@ -1,27 +1,34 @@
 using Documents.Models;
 using Documents.Services;
 using DocumentsAPI.Repositories;
+using DocumentsAPI.Services;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using System.Net;
+using System.Reflection.Metadata;
+using Document = Documents.Models.Document;
 
 namespace Documents.Controllers
 {
+#if DEVELOPMENT == false
     [Authorize]
+#endif
     [ApiController]
     public class DocumentsController : ControllerBase
     {
         private readonly ILogger<DocumentsController> _logger;
-        private readonly IConfiguration _config ;
+        private readonly IConfiguration _config;
         private readonly IDocumentRepository _documentsRepository;
         private readonly IDocumentsService _documentsService;
+        private readonly IFoldersService _foldersService;
 
-        public DocumentsController(IConfiguration config, IDocumentRepository documentsRepository, ILogger<DocumentsController> logger,IDocumentsService documentsService)
+        public DocumentsController(IConfiguration config, IDocumentRepository documentsRepository, IFoldersService foldersService,ILogger<DocumentsController> logger, IDocumentsService documentsService)
         {
             _config = config;
             _documentsRepository = documentsRepository;
             _logger = logger;
             _documentsService = documentsService;
+            _foldersService = foldersService;
         }
 
         // POST: Upload document(s)
@@ -49,6 +56,9 @@ namespace Documents.Controllers
                 this.HttpContext.Response.StatusCode = (int)HttpStatusCode.MultiStatus;
                 return documents;
             }
+
+            if (folderId != null) await _foldersService.UpdateFolderHasDocuments((Guid)folderId, true);
+
             // all documents ok
             return Ok(documents);
         }
@@ -61,7 +71,7 @@ namespace Documents.Controllers
         [ProducesResponseType((int)HttpStatusCode.InternalServerError)]
         public async Task<ActionResult<IEnumerable<Document>>> GetDocumentsAsync(Guid archiveId, [FromQuery] Guid? folderId = null, [FromQuery] bool includeDeleted = false)
         {
-            return Ok(await _documentsService.GetDocumentsAsync(archiveId, 100, includeDeleted));
+            return Ok(await _documentsService.GetDocumentsAsync(archiveId, 100, folderId, includeDeleted));
         }
 
         // GET: Download document
@@ -85,7 +95,14 @@ namespace Documents.Controllers
         [ProducesResponseType((int)HttpStatusCode.InternalServerError)]
         public async Task<IActionResult> DeleteAsync(Guid archiveId, Guid documentId)
         {
+            var document = await _documentsService.GetDocumentByIdAsync(archiveId, documentId);
+            if (document == null || document == new Document()) { throw new Exception($"Document with ID {documentId} not found"); }
+            var docsInFolder = await _documentsService.GetDocumentsAsync(archiveId, 100, document.FolderId, false);
             await _documentsService.DeleteAsync(archiveId, documentId);
+            // If only one document remains in the folder, mark the folder as empty after deleting the document
+            if (docsInFolder.Count() == 1)
+                await _foldersService.UpdateFolderHasDocuments((Guid)document.FolderId, false);
+
             return NoContent();
         }
 
@@ -98,6 +115,10 @@ namespace Documents.Controllers
         public async Task<IActionResult> UndeleteAsync(Guid archiveId, Guid documentId)
         {
             await _documentsService.UndeleteAsync(archiveId, documentId);
+            var document = await _documentsService.GetDocumentByIdAsync(archiveId, documentId);
+            await _foldersService.UpdateFolderHasDocuments((Guid)document.FolderId, true);
+
+
             return NoContent();
         }
 
@@ -119,9 +140,11 @@ namespace Documents.Controllers
         [ProducesResponseType((int)HttpStatusCode.NoContent)]
         [ProducesResponseType((int)HttpStatusCode.NotFound)]
         [ProducesResponseType((int)HttpStatusCode.InternalServerError)]
-        public async Task<IActionResult> CopyAsync(Guid archiveId, Guid documentId, [FromForm] string documentName)
+        public async Task<IActionResult> CopyAsync(Guid archiveId, Guid documentId, [FromForm] string documentName, [FromForm] Guid? folderId = null)
         {
-            await _documentsService.CopyAsync(archiveId, documentId, documentName);
+            await _documentsService.CopyAsync(archiveId, documentId, documentName, folderId);
+            if (folderId != null) await _foldersService.UpdateFolderHasDocuments((Guid)folderId, true);
+
             return NoContent();
         }
